fix(onboard): keep Colima gateway guard out of entrypoint
